Move AST types to separate module.
